Rename update header
